Backend/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
python-decouple
pydantic
pydantic[email]
passlib[bcrypt]
sqlalchemy
python-multipart
